+++ core/pathname.ts
@@ -2,7 +2,7 @@
 export function getPathname(path: string, indexFiles: string[]): string {
 	// TODO
 	let m;
-	const [dirname, basename] = (m = path.match(/^(.*?\/)([^\/]*)$/)) ? [m[1], m[2]] : ["", path];
+	const [dirname, basename] = (m = path.match(/^(.*?\/)([^\/]*)$/)) ? [m[1], m[2]] : ["./", path];
 	if (indexFiles.some((indexFile) => indexFile == basename)) {
 		return dirname;
 	}